[REFACTOR] Ensure user_pass not included in export.
Although user_pass would have been hashed, it still made little to no
sense including it in the export.

A common use case is to export from one environment to impose on to
another and if not careful this would include the hashed password as the
users password and thus set as the password (it would get re-hashed of
course). Storing passwords in text files is generally frowned upon, so
make sure we don't incldue it in the export.