[BUG] Fix subdomain install adding extra site.
When using a subdomain install, an extra site was being created when the
site key was an empty string, e.g.

example.com
.example.com
sub.example.com

When desired outcome is:

example.com
sub.example.com

This PR updates it so that for subdomain installs, if the site key is an
empty string it will no longer create a site with an empty prefix.